+++ src/indexPackage.ts
@@ -1,1 +1,2 @@
 export { default, ITreeMapProps, ColorModel } from "./components/TreeMap";
+export { NumberOfChildrenPlacement } from "./components/Node";